requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
supabase==2.17.0
facebook-business==18.0.0
pandas==2.1.3
python-dateutil==2.8.2
Werkzeug==2.3.7
google-ads==24.1.0

# Dependências de segurança e produção
gunicorn==21.2.0
supervisor==4.2.5
psutil==5.9.6
